+++ bpdm-gate-api/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/api/model/SharableRelationType.kt
@@ -21,5 +21,6 @@ package org.eclipse.tractusx.bpdm.gate.api.model
 
 enum class SharableRelationType {
     IsAlternativeHeadquarterFor,
-    IsManagedBy
+    IsManagedBy,
+    IsOwnedBy
 }